+++ backend/internal/server/apis/sports/stats/goals.go
@@ -2,14 +2,13 @@ package stats
 
 import (
 	"github.com/gofiber/fiber/v2"
+	"github.com/jak103/powerplay/internal/db"
+	"github.com/jak103/powerplay/internal/models"
 	"github.com/jak103/powerplay/internal/server/apis"
 	"github.com/jak103/powerplay/internal/server/services/auth"
-	"github.com/jak103/powerplay/internal/utils/responder"
-	"github.com/jak103/powerplay/internal/models"
-	"github.com/jak103/powerplay/internal/db"
 	"github.com/jak103/powerplay/internal/utils/locals"
+	"github.com/jak103/powerplay/internal/utils/responder"
 )
-
 
 func init() {
 	apis.RegisterHandler(fiber.MethodPost, "/goals", auth.Public, postGoalsHandler)
@@ -21,9 +20,9 @@ func postGoalsHandler(c *fiber.Ctx) error {
 	log.Debug("body: %q", c.Request().Body())
 	goalPostRequest := new(models.Goal)
 	err := c.BodyParser(goalPostRequest)
-	
+
 	// If valid structure in post request, continue on
-	if err != nil{
+	if err != nil {
 		log.WithErr(err).Error("Failed to parse Goal POST request.")
 		return err
 	}
@@ -31,8 +30,8 @@ func postGoalsHandler(c *fiber.Ctx) error {
 	// Connect to database and insert goal
 	db := db.GetSession(c)
 	record, err := db.SaveGoal(goalPostRequest)
-	
-	if err != nil{
+
+	if err != nil {
 		log.WithErr(err).Alert("Failed to parse goal request payload")
 		return responder.InternalServerError(c)
 	}
@@ -40,7 +39,7 @@ func postGoalsHandler(c *fiber.Ctx) error {
 	if record == nil {
 		return responder.BadRequest(c, "Could not post goal into database")
 	}
-	
+
 	return responder.Ok(c)
 
 }
