+++ static/oas/v1/openapi.yml
@@ -13,7 +13,7 @@ paths:
   /auth:
     $ref: "./auth/auth.yml#/paths/auth"
   /goals:
-    $ref: "./stats/goals.yml#/paths/goals"
+    $ref: "./sports/stats/goals.yml#/paths/goals"
   /leagues:
     $ref: "./sports/components/leagues.yml#/paths/leagues"
   /logos:
@@ -21,15 +21,15 @@ paths:
   /logos/{id}:
     $ref: "./sports/components/logos.yml#/paths/~1logos~1%7Bid%7D"
   /penalties:
-    $ref: "./stats/penalties.yml#/paths/penalties"
+    $ref: "./sports/stats/penalties.yml#/paths/penalties"
   /penaltyTypes:
-    $ref: "./stats/penalties.yml#/paths/penaltyTypes"
+    $ref: "./sports/stats/penalties.yml#/paths/penaltyTypes"
   /rosters:
     $ref: "./sports/components/rosters.yml#/paths/rosters"
   /seasons:
     $ref: "./sports/components/seasons.yml#/paths/seasons"
   /shotsongoal:
-    $ref: "./stats/shotsongoal.yml#/paths/shotsOnGoal"
+    $ref: "./sports/stats/shotsongoal.yml#/paths/shotsOnGoal"
   /teams:
     $ref: "./sports/components/teams.yml#/paths/~1teams"
   /teams/{id}:
@@ -37,7 +37,7 @@ paths:
   /user:
     $ref: "./users/users.yml#/paths/user"
   /venues:
-    $ref: "./stats/venues.yml#/paths/venues"
+    $ref: "./sports/stats/venues.yml#/paths/venues"
 
 components:
   securitySchemes:
