+++ static/oas/v1/sports/stats/goals.yml
