+++ static/oas/v1/sports/stats/penalties.yml
